alerts and bug fixes

Closes RENATUS-629, RENATUS-615, and RENATUS-613

See merge request 75f/android/renatus!298

+++ haystack/src/main/java/a75f/io/api/haystack/sync/HisSyncHandler.java
@@ -186,6 +186,10 @@ public class HisSyncHandler
     
         ArrayList<HashMap> devices = hayStack.readAll("device");
         for (HashMap device : devices) {
+            if (device.get("ccu") != null) {
+                continue;
+            }
+            CcuLog.d(TAG," sendHisToInflux device "+device.get("dis"));
             ArrayList<HashMap> points = hayStack.readAll("point and his and deviceRef == \""+device.get("id")+"\"");
             if (CCUHsApi.getInstance().getGUID(device.get("id").toString()) == null) {
                 entitySyncRequired = true;
@@ -208,7 +212,7 @@ public class HisSyncHandler
                     continue;
                 
                 }
-                ArrayList<HisItem> hisItems = (ArrayList<HisItem>) hayStack.tagsDb.getUnSyncedHisItems(HRef.copy(pointID));
+                /*ArrayList<HisItem> hisItems = (ArrayList<HisItem>) hayStack.tagsDb.getUnSyncedHisItems(HRef.copy(pointID));
                 if (hisItems.size() == 0) {
                     continue;
                 }
@@ -221,6 +225,15 @@ public class HisSyncHandler
                     item.setSyncStatus(true);
                 }
                 hayStack.tagsDb.setHisItemSyncStatus(hisItems);
+                hayStack.tagsDb.removeHisItems(HRef.copy(pointID));*/
+    
+                HisItem hisVal = hayStack.tagsDb.getLastHisItem(HRef.copy(pointID));
+                if (hisVal == null) {
+                    CcuLog.d(TAG, "His val not found : "+m.get("dis"));
+                    continue;
+                }
+                tsData.put( pointGUID.replace("@",""), String.valueOf(hisVal.getVal()));
+    
                 hayStack.tagsDb.removeHisItems(HRef.copy(pointID));
             
             }
